flask-server/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
effdet==0.4.1
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.0
executing==2.1.0
faiss-cpu==1.9.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.4.1
fsspec==2024.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.8.1
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
iopath==0.1.10
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.33
langchain-google-genai==2.0.1
langchain-text-splitters==0.3.0
langdetect==1.0.9
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-checkpoint-sqlite==2.0.1
langgraph-sdk==0.1.51
langsmith==0.3.4
layoutparser==0.3.4
libmagic==1.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.0
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.0
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.20.0
pikepdf==9.4.1
pillow==10.4.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pypandoc==1.15
pyparsing==3.1.4
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyreadline3==3.5.4
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.17
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
timm==1.0.11
tokenizers==0.20.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.5
unstructured-client==0.27.0
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.2
yarl==1.13.1
zstandard==0.23.0
